+++ install.ps1
@@ -33,13 +33,13 @@
     .\install.ps1
 
 .EXAMPLE
-    .\install.ps1 -InstallPath 'D:\Tools\AppPackager' -Version 2026.09.21.0091
+    .\install.ps1 -InstallPath 'D:\Tools\AppPackager' -Version 2026.09.21.0092
 
 .NOTES
     ScriptName : install.ps1
     Purpose    : Bootstrap install / update for AppPackager
     Owner      : CM Engineering
-    Version    : 2026.09.21.0091
+    Version    : 2026.09.21.0092
 #>
 
 [CmdletBinding()]